+++ FindPanel/OgreTextFindComponents/NSTextView/OgreTextViewUndoer.h
@@ -18,7 +18,7 @@
 {
     NSRange             *_rangeArray;
     NSMutableArray      *_attributedStringArray;
-    unsigned            _count, _tail;
+    NSUInteger          _count, _tail;
 }
 - (id)initWithCapacity:(unsigned)aCapacity;
 - (void)addRange:(NSRange)aRange attributedString:(NSAttributedString*)anAttributedString;
